+++ docs/CHANGELOG.rst
@@ -29,6 +29,8 @@ Changed
   ``resolwebio/chipseq:4.1.3`` Docker image
 - Use ``resolwebio/sra-tools:1.0.0`` Docker image in processes
   ``import-sra``, ``import-sra-single`` and ``import-sra-paired``
+- Optionally use sra-tools ``prefetch`` command when downloading and
+  converting SRA files to FASTQ format
 
 Fixed
 -----